Replace `iter_measurements` with `iter_digests`
When we get a measurement log from the RoT, each measurement includes an
identifier for the algorithm used to generate the measurement value (a
digest). When appraising this log we must ensure not only that the
digest values therein match known / trusted measurements byte for byte,
but also that they were generated by the same algorithm. If we don't
ensure that we're comparing digests of the same type then we risk the
(very unlikely) possibility that a different algorithm (and likely
input) was used to generate an identical digest output.

The previous implementation of `iter_measurements` omits this context.
This commit changes this function to iterate over the `Digest` type
which allows consumers access to both the digest value and the digest
algorithm.